+++ src/RTT_Format_Reader/SideData.cc
@@ -1,21 +1,20 @@
 //----------------------------------*-C++-*--------------------------------//
-/*! 
+/*!
  * \file   RTT_Format_Reader/SideData.cc
  * \author B.T. Adams
  * \date   Wed Jun 7 10:33:26 2000
  * \brief  Implementation file for RTT_Format_Reader/SideData class.
  * \note   Copyright (C) 2016-2018 Los Alamos National Security, LLC.
- *         All rights reserved.
- */
-//---------------------------------------------------------------------------//
-
+ *         All rights reserved. */
 //---------------------------------------------------------------------------//
 
 #include "SideData.hh"
 
 namespace rtt_RTT_Format_Reader {
+
+//----------------------------------------------------------------------------//
 /*!
- * \brief Parses the side_data block data from the mesh file via calls to 
+ * \brief Parses the side_data block data from the mesh file via calls to
  *        private member functions.
  * \param meshfile Mesh file name.
  */
@@ -25,6 +24,8 @@ void SideData::readSideData(ifstream &meshfile) {
     readData(meshfile);
   readEndKeyword(meshfile);
 }
+
+//----------------------------------------------------------------------------//
 /*!
  * \brief Reads and validates the side_data block keyword.
  * \param meshfile Mesh file name.
@@ -36,22 +37,26 @@ void SideData::readKeyword(ifstream &meshfile) {
   Insist(dummyString == "sidedat", "Invalid mesh file: sidedat block missing");
   std::getline(meshfile, dummyString);
 }
+
+//----------------------------------------------------------------------------//
 /*!
  * \brief Reads and validates the side data block data.
  * \param meshfile Mesh file name.
  */
 void SideData::readData(ifstream &meshfile) {
   string dummyString;
-  int sideNum;
+  size_t sideNum;
 
-  for (int i = 0; i < dims.get_nsides(); ++i) {
+  for (size_t i = 0; i < dims.get_nsides(); ++i) {
     meshfile >> sideNum;
     Insist(sideNum == i + 1, "Invalid mesh file: side data index out of order");
-    for (int j = 0; j < dims.get_nside_data(); ++j)
+    for (size_t j = 0; j < dims.get_nside_data(); ++j)
       meshfile >> data[i][j];
     std::getline(meshfile, dummyString);
   }
 }
+
+//----------------------------------------------------------------------------//
 /*!
  * \brief Reads and validates the end_sidedat block keyword.
  * \param meshfile Mesh file name.
